clearly

+++ include/Canvas.hpp
@@ -9,35 +9,35 @@
 
 class Canvas {
   public:
-    virtual size_t get_width() const = 0;
-    virtual size_t get_height() const = 0;
-    virtual void set_pixel(size_t row, size_t col, const sRGB &color) = 0;
-    virtual sRGB get_pixel(size_t row, size_t col) const = 0;
+    virtual int get_width() const = 0;
+    virtual int get_height() const = 0;
+    virtual void set_pixel(int row, int col, const sRGB &color) = 0;
+    virtual sRGB get_pixel(int row, int col) const = 0;
 
     virtual ~Canvas() = default;
 };
 
 class SFML_Canvas : public Canvas {
   public:
-    SFML_Canvas(size_t width, size_t height) : width(width), height(height), pixels(sf::PrimitiveType::Points, width * height) {
-        for (size_t row = 0; row < height; ++row)
-            for (size_t col = 0; col < width; ++col)
+    SFML_Canvas(int width, int height) : width(width), height(height), pixels(sf::PrimitiveType::Points, width * height) {
+        for (int row = 0; row < height; ++row)
+            for (int col = 0; col < width; ++col)
                 pixels[row * width + col].position = {static_cast<float>(col), static_cast<float>(row + 1)};
     }
 
-    size_t get_width() const override {
+    int get_width() const override {
         return width;
     }
 
-    size_t get_height() const override {
+    int get_height() const override {
         return height;
     }
 
-    void set_pixel(size_t row, size_t col, const sRGB &color) override {
+    void set_pixel(int row, int col, const sRGB &color) override {
         pixels[row * width + col].color = sf::Color(color.r, color.g, color.b);
     };
 
-    sRGB get_pixel(size_t row, size_t col) const override {
+    sRGB get_pixel(int row, int col) const override {
         sf::Color color = pixels[row * width + col].color;
         return {color.r, color.g, color.b};
     }
@@ -45,32 +45,32 @@ class SFML_Canvas : public Canvas {
     sf::VertexArray pixels;
 
   private:
-    size_t width;
-    size_t height;
+    int width;
+    int height;
 };
 
 class SimpleCanvas : public Canvas {
   public:
-    SimpleCanvas(size_t width, size_t height) : width(width), height(height), pixels(height, std::vector<sRGB>(width)) {}
+    SimpleCanvas(int width, int height) : width(width), height(height), pixels(height, std::vector<sRGB>(width)) {}
 
-    size_t get_width() const override {
+    int get_width() const override {
         return width;
     }
 
-    size_t get_height() const override {
+    int get_height() const override {
         return height;
     }
 
-    void set_pixel(size_t row, size_t col, const sRGB &color) override {
+    void set_pixel(int row, int col, const sRGB &color) override {
         pixels[row][col] = color;
     };
 
-    sRGB get_pixel(size_t row, size_t col) const override {
+    sRGB get_pixel(int row, int col) const override {
         return pixels[row][col];
     }
 
   private:
-    size_t width;
-    size_t height;
+    int width;
+    int height;
     std::vector<std::vector<sRGB>> pixels;
 };
